llamaindex_demo/requirements.txt
accelerate==0.31.0
faiss-cpu==1.8
gradio==4.26.0
llama-index==0.10.43
regex==2024.5.15
tiktoken==0.7.0
torch==2.3.1
tree-sitter<0.22.0
tree-sitter-languages==1.10.2
tqdm==4.66.4
transformers==4.39.0
zhipuai~=2.0
